rtl8723au: Fix Smatch errors in core/rtw_security.c
Smatch reports the following:

  CHECK   /home/finger/rtl8723au/core/rtw_security.c
/home/finger/rtl8723au/core/rtw_security.c:199 rtw_wep_encrypt() warn: potential memory corrupting cast 8 vs 4 bytes
/home/finger/rtl8723au/core/rtw_security.c:209 rtw_wep_encrypt() warn: potential memory corrupting cast 8 vs 4 bytes
/home/finger/rtl8723au/core/rtw_security.c:261 rtw_wep_decrypt() warn: potential memory corrupting cast 8 vs 4 bytes

These are caused by an unsigned long cast. That is correct for a 32-bit driver, but it
is an error for a 64-bit version.

+++ core/rtw_security.c
@@ -196,7 +196,7 @@ _func_enter_;
 
 				length=pattrib->last_txcmdsz-pattrib->hdrlen-pattrib->iv_len- pattrib->icv_len;
 
-				*((unsigned long *)crc)=cpu_to_le32(getcrc32(payload,length));
+				*((u32 *)crc)=cpu_to_le32(getcrc32(payload,length));
 
 				arcfour_init(&mycontext, wepkey,3+keylength);
 				arcfour_encrypt(&mycontext, payload, payload, length);
@@ -206,7 +206,7 @@ _func_enter_;
 			else
 			{
 			length=pxmitpriv->frag_len-pattrib->hdrlen-pattrib->iv_len-pattrib->icv_len ;
-				*((unsigned long *)crc)=cpu_to_le32(getcrc32(payload,length));
+				*((u32 *)crc)=cpu_to_le32(getcrc32(payload,length));
 				arcfour_init(&mycontext, wepkey,3+keylength);
 				arcfour_encrypt(&mycontext, payload, payload, length);
 				arcfour_encrypt(&mycontext, payload+length, crc, 4);
@@ -258,7 +258,7 @@ _func_enter_;
 		arcfour_encrypt(&mycontext, payload, payload,  length);
 
 		/* calculate icv and compare the icv */
-		*((unsigned long *)crc)=le32_to_cpu(getcrc32(payload,length-4));
+		*((u32 *)crc)=le32_to_cpu(getcrc32(payload,length-4));
 
 		if(crc[3]!=payload[length-1] || crc[2]!=payload[length-2] || crc[1]!=payload[length-3] || crc[0]!=payload[length-4])
 		{
